Rename from cudnn_conv_padding_legalization to gpu_conv_padding_legalization
